+++ BraintreeCore/src/main/java/com/braintreepayments/api/AuthorizationLoader.java
@@ -6,7 +6,6 @@ import androidx.annotation.Nullable;
 class AuthorizationLoader {
 
     private Authorization authorization;
-    private boolean invalidateNextAuthorizationRequest = false;
     private final ClientTokenProvider clientTokenProvider;
 
     AuthorizationLoader(@Nullable String initialAuthString, @Nullable ClientTokenProvider clientTokenProvider) {
@@ -17,10 +16,6 @@ class AuthorizationLoader {
     }
 
     void loadAuthorization(@NonNull final AuthorizationCallback callback) {
-        if (invalidateNextAuthorizationRequest && clientTokenProvider != null) {
-            authorization = null;
-        }
-
         if (authorization != null) {
             callback.onAuthorizationResult(authorization, null);
         } else if (clientTokenProvider != null) {
@@ -50,6 +45,8 @@ class AuthorizationLoader {
     }
 
     void invalidateClientToken() {
-        invalidateNextAuthorizationRequest = true;
+        if (clientTokenProvider != null) {
+            authorization = null;
+        }
     }
 }